\name{kkjokes}
\alias{kkjokes}
\docType{data}
\title{
List with Knock-Knock Jokes
}
\description{
List with Knock-Knock Jokes.
}
\usage{data(kkjokes)}
\format{
  List with Knock-Knock Jokes. Elements are of \code{"KnockKnockJoke"}
  class.
}
\source{
Various folks around me, kid's books and Internet.
}
\examples{
data(kkjokes)
kkjokes[[1]]
}
\keyword{datasets}
