\name{KnockKnockJokes-package}
\alias{KnockKnockJokes-package}
\alias{KnockKnockJokes}
\docType{package}
\title{
Knock-Knock Jokes
}
\description{
An S4 exercise for Knock-Knock Joke lovers.
}
\details{
Main function: \code{\link{KnockKnock}}
}
\author{
Bence, Lilla, and Peter Solymos

Maintainer: Peter Solymos <psolymos@gmail.com>
}
\keyword{ package }
\examples{
KnockKnock()
}
